Fix IBAN sanitizing
The IBAN was not sanitized correctly at all. For instance, spaces
were allowed among other things. $fields is not an editable array, the
values are just for validation and are ignored when saving changes to the
database.

Even worse, the IBAN constructor was called without the IBAN value as
argument. This caused the MachineFormat function to do nothing.